backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
langgraph
langchain-core
langchain-anthropic
python-telegram-bot
python-dotenv
httpx
SQLAlchemy
langgraph-checkpoint-sqlite
